Refactor http controllers
