+++ services/radio-manager-backend/src/http_server/handlers/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod internal_radio_streamer;
 pub(crate) mod public_schedule;
 pub(crate) mod public_streams;
+pub(crate) mod user_audio_stream;
 pub(crate) mod user_audio_tracks;
 pub(crate) mod user_audio_tracks_v2;
 pub(crate) mod user_stream_control;
