+++ services/radio-manager-backend/src/http_server/handlers/user_audio_tracks.rs
@@ -1,11 +1,7 @@
-use crate::config::Config;
 use crate::data_structures::{
     SortingColumn, SortingOrder, StreamId, TrackId, UserId, DEFAULT_TRACKS_PER_REQUEST,
 };
 use crate::http_server::response::Response;
-use crate::services::ffmpeg_service::{
-    transcode_audio_file, AudioChannels, AudioCodec, AudioContainer, TranscodeAudioFileFormat,
-};
 use crate::storage::db::repositories::streams::{
     get_single_stream_by_id, get_user_streams_having_track,
 };
@@ -15,16 +11,13 @@ use crate::storage::db::repositories::user_stream_tracks::{
 use crate::storage::db::repositories::user_tracks::{
     delete_user_track, get_single_user_track, get_user_tracks, GetUserTracksParams,
 };
-use crate::storage::db::row_utils::GetFilePath;
 use crate::storage::fs::utils::GetPath;
 use crate::storage::fs::FileSystem;
 use crate::utils::TeeResultUtils;
 use crate::MySqlClient;
 use actix_web::web::{Data, Form, Path, Query};
 use actix_web::{web, HttpResponse};
-use futures::channel::mpsc;
 use serde::Deserialize;
-use std::time::Duration;
 use tracing::error;
 
 #[derive(Deserialize)]
@@ -273,69 +266,4 @@ pub(crate) async fn delete_audio_track<FS: FileSystem>(
     connection.commit().await?;
 
     Ok(HttpResponse::Ok().finish())
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct TranscodeAudioTrackQuery {
-    #[serde(default)]
-    #[serde(with = "serde_millis")]
-    pub(crate) initial_position: Duration,
-}
-
-pub(crate) async fn transcode_audio_track(
-    user_id: UserId,
-    path: Path<TrackId>,
-    mysql_client: Data<MySqlClient>,
-    config: Data<Config>,
-    json: Query<TranscodeAudioTrackQuery>,
-) -> Response {
-    let mut connection = mysql_client.connection().await?;
-
-    let track_id = path.into_inner();
-    let track_row = match get_single_user_track(&mut connection, &track_id)
-        .await
-        .tee_err(|error| error!(?error, "Unable to get user track from database"))?
-    {
-        Some(track_row) => track_row,
-        None => return Ok(HttpResponse::NotFound().finish()),
-    };
-
-    if track_row.track.uid != user_id {
-        return Ok(HttpResponse::Forbidden().finish());
-    }
-
-    let source_path = track_row.get_file_path();
-    let source_url = format!("{}audio/{}", config.file_server_endpoint, source_path);
-
-    let (response_tx, response_rx) = mpsc::channel(32);
-
-    actix_rt::spawn({
-        async move {
-            if let Err(error) = transcode_audio_file(
-                &source_url,
-                response_tx,
-                json.initial_position,
-                TranscodeAudioFileFormat {
-                    container: AudioContainer::Adts,
-                    codec: AudioCodec::Aac,
-                    channels: AudioChannels::Stereo,
-                    bitrate: 256_000,
-                    sampling_rate: 48_000,
-                },
-            )
-            .await
-            {
-                error!(?error, "Error occurred while transcoding audio");
-            }
-        }
-    });
-
-    use futures::StreamExt;
-
-    let mut response = HttpResponse::Ok();
-
-    response.content_type("audio/aac").force_close();
-
-    Ok(response.streaming::<_, actix_web::Error>(response_rx.map(Ok)))
 }
